Simplified alternative - Pass song query directly
Alternative usage - not using prompt as in the index2.js version - same "simplified" search which returns the lyrics for the first song result for the query. (Identical to index2.js, just passing directly)

This takes the song value as a parameter when executing the script itself, i.e. when prompt cannot be used in a subprocess outside the terminal where we aren't able to keep track of a response.

Usage (from the project directory): 

node index3.js "SONG_QUERY_HERE" 

Example:

node index3.js "Eminem Rap God"  

The above will process and output the first result of the lyrics to the query "Eminem Rap God" in a single command instance.